src/modules/auth/useCases/validateUserUseCase/validateUserUseCase.ts
import { Injectable, UnauthorizedException } from '@nestjs/common';
import { compare } from 'bcrypt';
import { UserRepository } from 'src/modules/user/repositories/users.repository';
import { AuthValuesIncorrectException } from '../../exceptions/authValueIncorrectException';

interface ValidateUserRequest {
  name: string;
  password: string;
}

@Injectable()
export class ValidateUserUseCase {
  constructor(private userRepository: UserRepository) {}
  async execute({ name, password }: ValidateUserRequest) {
    const user = await this.userRepository.findByName(name);

    if (!user) throw new AuthValuesIncorrectException();

    const isPasswordMatched = await compare(password, user.password);

    if (!isPasswordMatched) throw new AuthValuesIncorrectException();

    return user;
  }
}
